--- packages/nextjs/components/headerMenuLinks.tsx
@@ -1,30 +1,0 @@
-import React from "react";
-import { BugAntIcon, MagnifyingGlassIcon, SparklesIcon } from "@heroicons/react/24/outline";
-
-interface HeaderMenuLink {
-  label: string;
-  href: string;
-  icon?: React.ReactNode;
-}
-
-export const headerMenuLinks: HeaderMenuLink[] = [
-  {
-    label: "Home",
-    href: "/",
-  },
-  {
-    label: "Debug Contracts",
-    href: "/debug",
-    icon: <BugAntIcon className="h-4 w-4" />,
-  },
-  {
-    label: "Example UI",
-    href: "/example-ui",
-    icon: <SparklesIcon className="h-4 w-4" />,
-  },
-  {
-    label: "Block Explorer",
-    href: "/blockexplorer",
-    icon: <MagnifyingGlassIcon className="h-4 w-4" />,
-  },
-];
